@@ -8,28 +8,29 @@ description = ""
 authors = ["Chris Wright <cwright@twigeducation.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.9"
-alembic = {extras = ["psycopg2"], version = "^1.6.5"}
-databases = {version = "^0.5.2", extras = ["postgresql"]}
-fastapi = "^0.68.0"
-gunicorn = "^20.1.0"
+python = "^3.12"
+alembic = {extras = ["psycopg2"], version = "^1.13.1"}
+databases = {version = "^0.9.0", extras = ["postgresql"]}
+fastapi = "^0.111.1"
+gunicorn = "^22.0.0"
 psycopg2 = "^2.9.1"
-pydantic = {extras = ["dotenv"], version = "^1.8.2"}
+pydantic = {version = "^2.6.4", extras = ["dotenv"]}
+pydantic-settings = "^2.2.1"
 python-jose = "^3.1.0"
-uvicorn = "^0.15.0"
+uvicorn = "^0.30.3"
 
-[tool.poetry.dev-dependencies]
-black = "^21.7b0"
-flake8 = "^3.9.2"
-httpx = "^0.19.0"
+[tool.poetry.group.dev.dependencies]
+black = "^24.3.0"
+flake8 = "^7.0.0"
+httpx = "^0.27.0"
 isort = "^5.9.2"
-pytest = "^6.2.4"
-pytest-asyncio = "^0.15.1"
-pytest-cov = "^2.12.1"
-pytest-forked = "^1.3.0"
-pytest-xdist = "^2.3.0"
+pytest = "^8.1.1"
+pytest-asyncio = "^0.23.6"
+pytest-cov = "^5.0.0"
+pytest-forked = "^1.6.0"
+pytest-xdist = "^3.5.0"
 requests = "^2.26.0"
-schemathesis = "^3.9.6"
+schemathesis = "^3.26.1"
 
 [build-system]
 requires = ["poetry>=1.1.7"]
