@@ -1,15 +1,14 @@
 from fastapi import FastAPI
 
 from .config import get_settings
-from .db import database_connect, database_disconnect
+from .db import lifespan
 from .routers import api_routers
 
 settings = get_settings()
 
 app = FastAPI(
     debug=settings.debug,
-    on_startup=[database_connect],
-    on_shutdown=[database_disconnect],
+    lifespan=lifespan,
 )
 
 for api_router in api_routers:
